ignore min and max aggregations in mql queries
