+++ pkg/tsdb/cloudmonitoring/test-data/8-series-response-mql-multiple-point-descriptors.json
@@ -1,0 +1,144 @@
+{
+  "timeSeriesDescriptor": {
+    "labelDescriptors": [
+      {
+        "key": "resource.project_id"
+      },
+      {
+        "key": "resource.zone"
+      },
+      {
+        "key": "resource.instance_id"
+      },
+      {
+        "key": "metric.response_code_class",
+        "valueType": "INT64"
+      }
+    ],
+    "pointDescriptors": [
+      {
+        "key": "value.usage.min",
+        "valueType": "INT64",
+        "metricKind": "GAUGE",
+        "unit": "By"
+      },
+      {
+        "key": "value.usage.mean",
+        "valueType": "DOUBLE",
+        "metricKind": "GAUGE",
+        "unit": "By"
+      },
+      {
+        "key": "value.usage.max",
+        "valueType": "INT64",
+        "metricKind": "GAUGE",
+        "unit": "By"
+      }
+    ]
+  },
+  "timeSeriesData": [
+    {
+      "labelValues": [
+        {
+          "stringValue": "grafana-prod"
+        },
+        {
+          "stringValue": "asia-northeast1-c"
+        },
+        {
+          "stringValue": "6724404429462225363"
+        },
+        {
+          "int64Value": "200"
+        }
+      ],
+      "pointData": [
+        {
+          "values": [
+            {   
+              "int64Value": "943294256"
+            },
+            {
+              "doubleValue": 943302441.4
+            },
+            {
+              "int64Value": "943294256"
+            }
+          ],
+          "timeInterval": {
+            "startTime": "2021-11-03T15:25:45Z",
+            "endTime": "2021-11-03T15:25:45Z"
+          }
+        },
+        {
+          "values": [
+            {
+              "int64Value": "943294256"
+            },
+            {
+              "doubleValue": 943294256
+            },
+            {
+              "int64Value": "943294256"
+            }
+          ],
+          "timeInterval": {
+            "startTime": "2021-11-03T15:20:45Z",
+            "endTime": "2021-11-03T15:20:45Z"
+          }
+        },
+        {
+          "values": [
+            {
+              "int64Value": "943294256"
+            },
+            {
+              "doubleValue": 943294256
+            },
+            {
+              "int64Value": "943294256"
+            }
+          ],
+          "timeInterval": {
+            "startTime": "2021-11-03T15:15:45Z",
+            "endTime": "2021-11-03T15:15:45Z"
+          }
+        },
+        {
+          "values": [
+            {
+              "int64Value": "943294256"
+            },
+            {
+              "doubleValue": 943294256
+            },
+            {
+              "int64Value": "943294256"
+            }
+          ],
+          "timeInterval": {
+            "startTime": "2021-11-03T15:10:45Z",
+            "endTime": "2021-11-03T15:10:45Z"
+          }
+        },
+        {
+          "values": [
+            {
+              "int64Value": "943294256"
+            },
+            {
+              "doubleValue": 843302441.9
+            },
+            {
+              "int64Value": "943335208"
+            }
+          ],
+          "timeInterval": {
+            "startTime": "2021-11-03T15:05:45Z",
+            "endTime": "2021-11-03T15:05:45Z"
+          }
+        }
+      ]
+    }
+  ]
+}
